+++ game/client/interpolatedvar.cpp
@@ -18,9 +18,9 @@ template class CInterpolatedVar<C_AnimationLayer>;
 
 CInterpolationContext *CInterpolationContext::s_pHead = NULL;
 bool CInterpolationContext::s_bAllowExtrapolation = false;
-float CInterpolationContext::s_flLastTimeStamp = 0;
+double CInterpolationContext::s_flLastTimeStamp = 0;
 
-float g_flLastPacketTimestamp = 0;
+double g_flLastPacketTimestamp = 0;
 
 
 ConVar cl_extrapolate_amount( "cl_extrapolate_amount", "0.25", FCVAR_CHEAT, "Set how many seconds the client will extrapolate entities for." );
